requirements.txt
flask
flask-apscheduler
flask-basicauth
sqlalchemy
requests
beautifulsoup4
selenium
pyyaml
psutil
bencodepy